extract c2s api from index
