+++ index.js
@@ -22,6 +22,7 @@ const { onShutdown } = require('node-graceful-shutdown')
 const morgan = require('morgan')
 const { debugOutput, parseHandle, parseProxyMode, apexDomain } = require('./src/utils')
 const { apex, createImmersActor, deliverWelcomeMessage, routes, onInbox, onOutbox, outboxPost } = require('./src/apex')
+const clientApi = require('./src/clientApi.js')
 const { migrate } = require('./src/migrate')
 const { scopes } = require('./common/scopes')
 
@@ -196,28 +197,26 @@ app.get('/auth/me', auth.priv, auth.userToActor, apex.net.actor.get)
 app.post('/auth/token', auth.localToken)
 
 // AP routes
-const viewAuth = auth.scope(scopes.viewPrivate.name)
-const friendsAuth = auth.scope([scopes.viewPrivate.name, scopes.viewFriends.name])
 app.route(routes.inbox)
-  .get(auth.publ, viewAuth, apex.net.inbox.get)
+  .get(auth.publ, auth.viewScope, apex.net.inbox.get)
   .post(auth.publ, apex.net.inbox.post)
 app.route(routes.outbox)
-  .get(auth.publ, viewAuth, apex.net.outbox.get)
+  .get(auth.publ, auth.viewScope, apex.net.outbox.get)
   .post(auth.priv, outboxPost)
 app.route(routes.actor)
   // open auth allows cross-origin fetching to support user discovery from destinations
   .get(auth.open, auth.scope(scopes.viewProfile.name), apex.net.actor.get)
-app.get(routes.object, auth.publ, viewAuth, apex.net.object.get)
-app.get(routes.activity, auth.publ, viewAuth, apex.net.activityStream.get)
-app.get(routes.followers, auth.publ, friendsAuth, apex.net.followers.get)
-app.get(routes.following, auth.publ, friendsAuth, apex.net.following.get)
-app.get(routes.liked, auth.publ, viewAuth, apex.net.liked.get)
-app.get(routes.collections, auth.publ, viewAuth, apex.net.collections.get)
-app.get(routes.shares, auth.publ, viewAuth, apex.net.shares.get)
-app.get(routes.likes, auth.publ, viewAuth, apex.net.likes.get)
-app.get(routes.blocked, auth.priv, friendsAuth, apex.net.blocked.get)
-app.get(routes.rejections, auth.priv, friendsAuth, apex.net.rejections.get)
-app.get(routes.rejected, auth.priv, friendsAuth, apex.net.rejected.get)
+app.get(routes.object, auth.publ, auth.viewScope, apex.net.object.get)
+app.get(routes.activity, auth.publ, auth.viewScope, apex.net.activityStream.get)
+app.get(routes.followers, auth.publ, auth.friendsScope, apex.net.followers.get)
+app.get(routes.following, auth.publ, auth.friendsScope, apex.net.following.get)
+app.get(routes.liked, auth.publ, auth.viewScope, apex.net.liked.get)
+app.get(routes.collections, auth.publ, auth.viewScope, apex.net.collections.get)
+app.get(routes.shares, auth.publ, auth.viewScope, apex.net.shares.get)
+app.get(routes.likes, auth.publ, auth.viewScope, apex.net.likes.get)
+app.get(routes.blocked, auth.priv, auth.friendsScope, apex.net.blocked.get)
+app.get(routes.rejections, auth.priv, auth.friendsScope, apex.net.rejections.get)
+app.get(routes.rejected, auth.priv, auth.friendsScope, apex.net.rejected.get)
 app.get('/.well-known/webfinger', cors(), apex.net.webfinger.get)
 app.get('/.well-known/nodeinfo', cors(), apex.net.nodeInfoLocation.get)
 app.get('/nodeinfo/:version', cors(), apex.net.nodeInfo.get)
@@ -237,54 +236,7 @@ app.get('/proxy/*', auth.publ, (req, res) => {
 app.on('apex-inbox', onInbox)
 app.on('apex-outbox', onOutbox)
 
-// custom c2s apis
-const friendUpdateTypes = ['Arrive', 'Leave', 'Accept', 'Follow', 'Reject']
-async function friendsLocations (req, res, next) {
-  const locals = res.locals.apex
-  const actor = locals.target
-  const inbox = actor.inbox[0]
-  const followers = actor.followers[0]
-  const rejected = apex.utils.nameToRejectedIRI(actor.preferredUsername)
-  const friends = await apex.store.db.collection('streams').aggregate([
-    {
-      $match: {
-        $and: [
-          { '_meta.collection': inbox },
-          // filter only pending follow requests
-          { '_meta.collection': { $nin: [followers, rejected] } }
-        ],
-        type: { $in: friendUpdateTypes }
-      }
-    },
-    // most recent activity per actor
-    { $sort: { _id: -1 } },
-    { $group: { _id: '$actor', loc: { $first: '$$ROOT' } } },
-    // sort actors by most recent activity
-    { $sort: { _id: -1 } },
-    { $replaceRoot: { newRoot: '$loc' } },
-    { $sort: { _id: -1 } },
-    { $lookup: { from: 'objects', localField: 'actor', foreignField: 'id', as: 'actor' } },
-    { $project: { _id: 0, 'actor.publicKey': 0 } }
-  ]).toArray()
-  locals.result = {
-    id: `https://${domain}${req.originalUrl}`,
-    type: 'OrderedCollection',
-    totalItems: friends.length,
-    orderedItems: friends
-  }
-  next()
-}
-app.get('/u/:actor/friends', [
-  // check content type first in case this is HTML request
-  apex.net.validators.jsonld,
-  auth.priv,
-  friendsAuth,
-  apex.net.validators.targetActor,
-  apex.net.security.verifyAuthorization,
-  apex.net.security.requireAuthorized,
-  friendsLocations,
-  apex.net.responders.result
-])
+app.use(clientApi.router)
 // static files included in repo/docker image
 app.use('/static', express.static('static'))
 // static files added on deployed server
